좋아요, 이전다음페이지 완료

+++ gasil/src/main/java/org/joonzis/controller/HeartController.java
@@ -22,41 +22,38 @@ public class HeartController {
 	private final HeartService service;
 	private final BoardService bservice;
 	
+	/*
+	 * @PostMapping("/like")
+	 * 
+	 * @ResponseBody public int likeConfirm(@RequestBody HeartVO vo) { // JSON으로 값이
+	 * 들어오기 때문에 @RequestBody와 커맨드 객체를 사용해서 JSON 타입을 객첼 변경
+	 * System.out.println("좋아요 기능 값을 가져오는지 확인 : " + vo);
+	 * 
+	 * service.insert(vo); return bservice.upHeart(vo.getBno()); }
+	 * 
+	 * @PostMapping("/unlike")
+	 * 
+	 * @ResponseBody public int unlikeConfirm(@RequestBody HeartVO vo) {
+	 * System.out.println("좋아요 취소 기능 값을 가져오는지 확인 : " + vo);
+	 * 
+	 * service.deleteLike(vo); return bservice.downHeart(vo.getBno()); }
+	 */
+	
+	
 	@PostMapping("/like")
 	@ResponseBody
-	public int likeConfirm(@RequestBody HeartVO vo) {
-		// JSON으로 값이 들어오기 때문에 @RequestBody와 커맨드 객체를 사용해서 JSON 타입을 객첼 변경 
+	public int heartConfirm(@RequestBody HeartVO vo) {
 		System.out.println("좋아요 기능 값을 가져오는지 확인 : " + vo);
-		
-		service.insert(vo);
-		return bservice.upHeart(vo.getBno());
-	}
-		@PostMapping("/unlike")
-		@ResponseBody
-		public int unlikeConfirm(@RequestBody HeartVO vo) {
-			System.out.println("좋아요 취소 기능 값을 가져오는지 확인 : " + vo);
-			
+		// 하트를 누른 상태면 result는 1, 아니면 0
+		int result = service.checkLike(vo);
+		if(result == 0) {
+			service.insert(vo);
+			bservice.upHeart(vo.getBno());
+		}else {
 			service.deleteLike(vo);
-			return bservice.downHeart(vo.getBno());
+			bservice.downHeart(vo.getBno());
 		}
+		return service.totalLike(vo.getBno());
 	}
 	
-	
-	/*
-	 * @ResponseBody
-	 * 
-	 * @PostMapping("/likeUp") public void likeUp(@RequestBody HeartVO vo) {
-	 * System.out.println(vo.getBno()); System.out.println(vo.getUser_id());
-	 * 
-	 * service.likeUp(vo); }
-	 * 
-	 * @ResponseBody
-	 * 
-	 * @PostMapping("/likeDown") public void likeDown(@RequestParam ("bno") long
-	 * bno, @RequestParam("user_id") String user_id) {
-	 * System.out.println("좋아요 취소 : " + bno); System.out.println("좋아요 취소 : " +
-	 * user_id);
-	 * 
-	 * HeartVO vo = new HeartVO(); vo.setBno(bno); vo.setUser_id(user_id);
-	 * service.likeDown(vo); }
-	 */
+}
